requirements.txt
torch
torchvision
streamlit==1.29.0
numpy
pillow
flask
